chore: move types to qlearning types file
